analytics: DRY analytics middlewares, reenable data events
