+++ frontend/scripts/index.js
@@ -2,7 +2,7 @@
 import { createStore, combineReducers, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
 import rangeTouch from 'rangetouch';
-import analyticsMiddleware from 'analytics/tool.analytics.middleware';
+import analyticsMiddleware from 'analytics/middleware';
 import { toolUrlStateMiddleware } from 'utils/stateURL';
 import router from './router/router';
 import routeSubscriber from './router/route-subscriber';
